backend/requirements.txt
Django==1.11.29
djangorestframework==3.11.0
psycopg2==2.8.4
drf-extensions==0.6.0
django-extensions==2.2.8
django-cors-headers==3.2.1

django-suit==0.2.28
gunicorn==20.0.4
